staging: omapdrm: Remove unnecessary memcpy
Removed the unnecessary copy of the memory page addresses when
programming the DMM/PAT and all support code for the lut copy.
The original intent was to have this code in place for
suspend/resume functionality w.r.t. DEVICE_OFF.

Performance analysis showed that the extra copy from uncached memory
led to a fairly hefty penalty when programming large 1D or 2D
buffers.  This can be implemented in a more efficient manner when we
actually have to support DEVICE_OFF suspend/resume operations.

Change-Id: I62487cbdbc90270bf4c542641129ad3ef52bd959
Signed-off-by: Andy Gross <andy.gross@ti.com>