OMAP5: EHCI:Fix EHCI crash during reboot
In case of system reboot triggered from application level to reboot the system,
the EHCI driver shutdown will be called from system shutdown process to shut
the EHCI dirver. As part of this shutdown process EHCI driver accesses some of
EHCI registers to shut the driver properly. This works well in case EHCI is in
active state before reboot is triggered from the system level, but however in
case of USB auto-suspend this causes crash as the clocks of EHCI have been
already disabled as part of auto-suspend.In our current system the USB auto
suspend is enabled by default for CPUIdle to work properly.

So fix is to check if the driver is in suspended state in case of
before shutting down the ECHI driver, then enable the clocks by
pm_runtime_get_sync.

Change-Id: I3eb292f5e240ded74f43bc7efdae927473482ba6
Reported-by: Moiz Sonasath <m-sonasath@ti.com>
Reported-by: Menon, Nishanth <nm@ti.com>
Signed-off-by: Mantesh Sarasetti <mantesh@ti.com>