rpmsg: omx: remove ion support
Remove ion support in rpmsg omx driver.

I kept ioctls register/unregister not to break current code which uses
those ioctls. The reason is that, the ion framework was migrated to use
dma_buf framework inside, so a ion fd is really a dma_buf fd. Therefore,
in the ion register api we can copy the fd as the handle and it will work
with the dma_buf code. That means that previous code unsing ion register/
unregister ioctls will still work.

OTOH, the PVR support was removed completely since there is no way to make
it work right now.

Change-Id: Ib9177b7ecd7e4ea6f46aefd18d0ad4a9a2336ea5
Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>